add Y1 to the function defintion of estimate.
Prior to this change Y1 was documented as functional arguement but not
part of the function definition.  R CMD check identified Y1 as a
undefined global variable as a result.  I suspect no errors were thrown
during testing because the any use of Y1 in a call would be captured by
the `...` argument.

This change add Y1 explicitly to the function arguments.